+++ features/sharding/core/src/main/java/org/apache/shardingsphere/sharding/route/engine/condition/generator/ConditionValueGenerator.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.sharding.route.engine.condition.generator;
 
-import org.apache.shardingsphere.sharding.route.engine.condition.Column;
+import org.apache.shardingsphere.infra.metadata.database.schema.HashColumn;
 import org.apache.shardingsphere.sharding.route.engine.condition.value.ShardingConditionValue;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.dml.expr.ExpressionSegment;
 import org.apache.shardingsphere.timeservice.core.rule.TimestampServiceRule;
@@ -41,5 +41,5 @@ public interface ConditionValueGenerator<T extends ExpressionSegment> {
      * @param timestampServiceRule time service rule
      * @return route value
      */
-    Optional<ShardingConditionValue> generate(T predicateRightValue, Column column, List<Object> params, TimestampServiceRule timestampServiceRule);
+    Optional<ShardingConditionValue> generate(T predicateRightValue, HashColumn column, List<Object> params, TimestampServiceRule timestampServiceRule);
 }