+++ kernel/transaction/type/xa/core/src/main/java/org/apache/shardingsphere/transaction/xa/jta/datasource/XATransactionDataSource.java
@@ -17,8 +17,8 @@
 
 package org.apache.shardingsphere.transaction.xa.jta.datasource;
 
-import org.apache.shardingsphere.infra.database.core.type.DatabaseType;
 import org.apache.shardingsphere.infra.database.core.spi.DatabaseTypedSPILoader;
+import org.apache.shardingsphere.infra.database.core.type.DatabaseType;
 import org.apache.shardingsphere.infra.util.reflection.ReflectionUtils;
 import org.apache.shardingsphere.transaction.xa.jta.connection.XAConnectionWrapper;
 import org.apache.shardingsphere.transaction.xa.jta.datasource.properties.XADataSourceDefinition;
@@ -36,10 +36,13 @@ import javax.transaction.Transaction;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedList;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * XA transaction data source.
@@ -48,7 +51,9 @@ public final class XATransactionDataSource implements AutoCloseable {
     
     private static final Set<String> CONTAINER_DATASOURCE_NAMES = new HashSet<>(Arrays.asList("AtomikosDataSourceBean", "BasicManagedDataSource"));
     
-    private final ThreadLocal<Map<Transaction, Connection>> enlistedTransactions = ThreadLocal.withInitial(HashMap::new);
+    private final ThreadLocal<Map<Transaction, Collection<Connection>>> enlistedTransactions = ThreadLocal.withInitial(HashMap::new);
+    
+    private final AtomicInteger uniqueName = new AtomicInteger();
     
     private final String resourceName;
     
@@ -84,25 +89,25 @@ public final class XATransactionDataSource implements AutoCloseable {
             return dataSource.getConnection();
         }
         Transaction transaction = xaTransactionManagerProvider.getTransactionManager().getTransaction();
-        if (!enlistedTransactions.get().containsKey(transaction)) {
-            Connection connection = dataSource.getConnection();
-            XAConnection xaConnection = xaConnectionWrapper.wrap(xaDataSource, connection);
-            transaction.enlistResource(new SingleXAResource(resourceName, xaConnection.getXAResource()));
-            transaction.registerSynchronization(new Synchronization() {
-                
-                @Override
-                public void beforeCompletion() {
-                    enlistedTransactions.get().remove(transaction);
-                }
-                
-                @Override
-                public void afterCompletion(final int status) {
-                    enlistedTransactions.get().clear();
-                }
-            });
-            enlistedTransactions.get().put(transaction, connection);
-        }
-        return enlistedTransactions.get().get(transaction);
+        Connection connection = dataSource.getConnection();
+        XAConnection xaConnection = xaConnectionWrapper.wrap(xaDataSource, connection);
+        transaction.enlistResource(new SingleXAResource(resourceName, String.valueOf(uniqueName.getAndIncrement()), xaConnection.getXAResource()));
+        transaction.registerSynchronization(new Synchronization() {
+            
+            @Override
+            public void beforeCompletion() {
+                enlistedTransactions.get().remove(transaction);
+                uniqueName.set(0);
+            }
+            
+            @Override
+            public void afterCompletion(final int status) {
+                enlistedTransactions.get().clear();
+            }
+        });
+        enlistedTransactions.get().computeIfAbsent(transaction, key -> new LinkedList<>());
+        enlistedTransactions.get().get(transaction).add(connection);
+        return connection;
     }
     
     @Override